brinquedos-por-assinatura-angola/backend-python/requirements.txt

flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
